filetocolormap

+++ ColorMapInput.c
@@ -42,7 +42,7 @@ uint8_t** FileToColorMap(char* colorfile, int* colorcount)
 
     uint8_t** carray = (uint8_t**) malloc (sizeof(uint8_t*));
     uint8_t* rgb[3];
-    uint8_t* r, g, b;
+    int* r, g, b;
     while(*colorcount < numcount) {
         if (fscanf(cfile, "%d %d %d", &r, &g, &b) != 3) {
             return NULL;
